Address review feedback on the update checker
