+++ src/ServerMainComponent.cpp
@@ -6,6 +6,7 @@
 #include "ServerMainComponent.hpp"
 
 #include "AlarmMaskAudio.h"
+#include "IndentedToggleButton.hpp"
 #include "JuceManagedWorkingSetCache.hpp"
 #include "Main.hpp"
 #include "ShortcutsWindow.hpp"
@@ -26,50 +27,6 @@
 #include <iomanip>
 #include <iterator>
 #include <sstream>
-
-namespace
-{
-	class IndentedToggleButton : public juce::Component
-	{
-	public:
-		IndentedToggleButton()
-		{
-			addAndMakeVisible(toggle);
-		}
-
-		juce::ToggleButton toggle;
-
-	private:
-		void moved() override
-		{
-			layout_toggle();
-		}
-
-		void resized() override
-		{
-			layout_toggle();
-		}
-
-		void layout_toggle()
-		{
-			int tickLeft = 90;
-			if (auto *alert = getParentComponent())
-			{
-				const int iconSize = juce::jmin(130, alert->getHeight() + 18);
-				const float centre = static_cast<float>(iconSize / -10 + iconSize / 2);
-				const float radius = iconSize * 0.5f;
-				const float y = static_cast<float>(juce::jlimit(getY() + 4, getY() + getHeight() - 4, static_cast<int>(centre)));
-				const float dy = y - centre;
-				const float circleRight = (std::abs(dy) < radius) ? centre + std::sqrt(radius * radius - dy * dy) : centre;
-				tickLeft = static_cast<int>(std::ceil(circleRight)) + 6;
-			}
-
-			const int tickInset = 4; // LookAndFeel_V4 draws the tick box this far inside the button
-			const int indent = juce::jmax(0, tickLeft - getX() - tickInset);
-			toggle.setBounds(indent, 0, juce::jmax(0, getWidth() - indent), getHeight());
-		}
-	};
-}
 
 ServerMainComponent::ServerMainComponent(
   std::shared_ptr<isobus::InternalControlFunction> serverControlFunction,
@@ -1921,97 +1878,9 @@ void ServerMainComponent::check_load_settings(std::shared_ptr<ValueTree> setting
 void ServerMainComponent::check_for_update(bool reportWhenUpToDate)
 {
 	auto checkStarted = updateChecker.start([safeThis = Component::SafePointer<ServerMainComponent>(this), reportWhenUpToDate](const UpdateChecker::Result &result) {
-		if (nullptr == safeThis.getComponent())
+		if (nullptr != safeThis.getComponent())
 		{
-			return; // The VT was closed before the check finished
-		}
-
-		safeThis->mCommandManager.commandStatusChanged();
-		const bool updateIsSkipped = result.updateAvailable && (safeThis->skippedUpdateVersion == result.latestVersion);
-
-		if (result.checkSucceeded && (!result.updateAvailable || updateIsSkipped))
-		{
-			safeThis->lastUpdateCheck = Time::getCurrentTime().toMilliseconds();
-			safeThis->save_settings();
-		}
-
-		if (result.updateAvailable)
-		{
-			isobus::CANStackLogger::info("Version " + result.latestVersion.toStdString() + " of this application is available at " + result.releaseUrl.toStdString());
-
-			// On startup we stay quiet about a version the user chose to skip. A manual check still
-			// shows the dialog with the box ticked, so they can un-skip it from there.
-			if (updateIsSkipped && !reportWhenUpToDate)
-			{
-				return;
-			}
-
-			auto skipToggle = std::make_shared<IndentedToggleButton>();
-			skipToggle->toggle.setButtonText("Don't notify me about version " + result.latestVersion + " again");
-			skipToggle->setSize(360, 24);
-			skipToggle->toggle.setToggleState(updateIsSkipped, NotificationType::dontSendNotification);
-
-			auto *updateBox = new AlertWindow("Update Available",
-			                                  "Version " + result.latestVersion + " is available. You are running " + String(ProjectInfo::versionString) + ".",
-			                                  MessageBoxIconType::InfoIcon);
-			updateBox->addCustomComponent(skipToggle.get());
-			updateBox->addButton("Download", 1, KeyPress(KeyPress::returnKey));
-			updateBox->addButton("Not Now", 0, KeyPress(KeyPress::escapeKey));
-			updateBox->enterModalState(true,
-			                           ModalCallbackFunction::create([safeThis, latestVersion = result.latestVersion, releaseUrl = result.releaseUrl, skipToggle](int choice) {
-				                           if (nullptr != safeThis.getComponent())
-				                           {
-					                           const String newSkippedVersion = skipToggle->toggle.getToggleState() ? latestVersion : String();
-
-					                           if (newSkippedVersion != safeThis->skippedUpdateVersion)
-					                           {
-						                           safeThis->skippedUpdateVersion = newSkippedVersion;
-						                           safeThis->lastUpdateCheck = 0;
-						                           safeThis->save_settings();
-					                           }
-				                           }
-
-				                           if (1 == choice)
-				                           {
-					                           URL(releaseUrl).launchInDefaultBrowser();
-				                           }
-			                           }),
-			                           true);
-		}
-		else if (result.checkSucceeded)
-		{
-			isobus::CANStackLogger::info("This application is up to date.");
-
-			if (reportWhenUpToDate)
-			{
-				AlertWindow::showAsync(MessageBoxOptions()
-				                         .withIconType(MessageBoxIconType::InfoIcon)
-				                         .withTitle("No Updates Available")
-				                         .withMessage("You are running the latest version, " + String(ProjectInfo::versionString) + ".")
-				                         .withButton("OK"),
-				                       nullptr);
-			}
-		}
-		else
-		{
-			auto failureMessage = std::string("Could not check GitHub for a newer version of this application.");
-
-			if (0 != result.statusCode)
-			{
-				failureMessage += " HTTP status code: " + std::to_string(result.statusCode) + ".";
-			}
-
-			isobus::CANStackLogger::warn(failureMessage);
-
-			if (reportWhenUpToDate)
-			{
-				AlertWindow::showAsync(MessageBoxOptions()
-				                         .withIconType(MessageBoxIconType::WarningIcon)
-				                         .withTitle("Update Check Failed")
-				                         .withMessage("Could not reach GitHub to check for a newer version.")
-				                         .withButton("OK"),
-				                       nullptr);
-			}
+			safeThis->on_update_check_complete(result, reportWhenUpToDate);
 		}
 	});
 
@@ -2033,6 +1902,97 @@ void ServerMainComponent::check_for_update(bool reportWhenUpToDate)
 			                         .withIconType(MessageBoxIconType::WarningIcon)
 			                         .withTitle("Update Check Failed")
 			                         .withMessage("Could not start the update check.")
+			                         .withButton("OK"),
+			                       nullptr);
+		}
+	}
+}
+
+void ServerMainComponent::on_update_check_complete(const UpdateChecker::Result &result, bool reportWhenUpToDate)
+{
+	mCommandManager.commandStatusChanged();
+	const bool updateIsSkipped = result.updateAvailable && (skippedUpdateVersion == result.latestVersion);
+
+	if (result.checkSucceeded && (!result.updateAvailable || updateIsSkipped))
+	{
+		lastUpdateCheck = Time::getCurrentTime().toMilliseconds();
+		save_settings();
+	}
+
+	if (result.updateAvailable)
+	{
+		isobus::CANStackLogger::info("Version " + result.latestVersion.toStdString() + " of this application is available at " + result.releaseUrl.toStdString());
+
+		// On startup we stay quiet about a version the user chose to skip. A manual check still
+		// shows the dialog with the box ticked, so they can un-skip it from there.
+		if (updateIsSkipped && !reportWhenUpToDate)
+		{
+			return;
+		}
+
+		auto skipToggle = std::make_shared<IndentedToggleButton>();
+		skipToggle->toggle.setButtonText("Don't notify me about version " + result.latestVersion + " again");
+		skipToggle->setSize(360, 24);
+		skipToggle->toggle.setToggleState(updateIsSkipped, NotificationType::dontSendNotification);
+
+		auto *updateBox = new AlertWindow("Update Available",
+		                                  "Version " + result.latestVersion + " is available. You are running " + String(ProjectInfo::versionString) + ".",
+		                                  MessageBoxIconType::InfoIcon);
+		updateBox->addCustomComponent(skipToggle.get());
+		updateBox->addButton("Download", 1, KeyPress(KeyPress::returnKey));
+		updateBox->addButton("Not Now", 0, KeyPress(KeyPress::escapeKey));
+		updateBox->enterModalState(true,
+		                           ModalCallbackFunction::create([safeThis = Component::SafePointer<ServerMainComponent>(this), latestVersion = result.latestVersion, releaseUrl = result.releaseUrl, skipToggle](int choice) {
+			                           if (nullptr != safeThis.getComponent())
+			                           {
+				                           const String newSkippedVersion = skipToggle->toggle.getToggleState() ? latestVersion : String();
+
+				                           if (newSkippedVersion != safeThis->skippedUpdateVersion)
+				                           {
+					                           safeThis->skippedUpdateVersion = newSkippedVersion;
+					                           safeThis->lastUpdateCheck = 0;
+					                           safeThis->save_settings();
+				                           }
+			                           }
+
+			                           if (1 == choice)
+			                           {
+				                           URL(releaseUrl).launchInDefaultBrowser();
+			                           }
+		                           }),
+		                           true);
+	}
+	else if (result.checkSucceeded)
+	{
+		isobus::CANStackLogger::info("This application is up to date.");
+
+		if (reportWhenUpToDate)
+		{
+			AlertWindow::showAsync(MessageBoxOptions()
+			                         .withIconType(MessageBoxIconType::InfoIcon)
+			                         .withTitle("No Updates Available")
+			                         .withMessage("You are running the latest version, " + String(ProjectInfo::versionString) + ".")
+			                         .withButton("OK"),
+			                       nullptr);
+		}
+	}
+	else
+	{
+		auto failureMessage = std::string("Could not check GitHub for a newer version of this application.");
+
+		if (0 != result.statusCode)
+		{
+			failureMessage += " HTTP status code: " + std::to_string(result.statusCode) + ".";
+		}
+
+		isobus::CANStackLogger::warn(failureMessage);
+
+		if (reportWhenUpToDate)
+		{
+			AlertWindow::showAsync(MessageBoxOptions()
+			                         .withIconType(MessageBoxIconType::WarningIcon)
+			                         .withTitle("Update Check Failed")
+			                         .withMessage("Could not reach GitHub to check for a newer version.")
 			                         .withButton("OK"),
 			                       nullptr);
 		}
